requirements.txt
requests
pytest
allure-pytest
PyYAML
Faker
loguru
pycryptodome
Jinja2
pydantic
pandas
openpyxl
autopep8
furl